searchBar + AboutUs page (css + js)
