+++ src/components/NavBar.css
@@ -1,7 +1,7 @@
 .logo {
-  height: 40px; /* Adjust height as needed */
+  height: 40px; 
   width: auto; 
-  margin-right: 0px; /* Add margin for spacing */
+  margin-right: 0px; 
   margin-bottom: 10px;
 }
 .navbar {
@@ -149,130 +149,4 @@
   font-size: 1rem;
 
   border: 10px;
-}
-/* Search container styles */
-.search-container {
-  position: relative;
-  display: flex;
-  align-items: center;
-
-}
-
-.search-button {
-  position: absolute;
-  left: 0;
-  width: 50px;
-  height: 50px;
-  border: none;
-  background: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  transition: color 0.3s ease, transform 0.2s ease, box-shadow 0.2s ease;
-  color: #a67c68;
-  box-shadow: none;
-  border-radius: 50%; /* Make it circular */
-}
-
-.search-button:hover,
-.search-button:focus {
-  color: #6e2020;
-  transform: scale(1.1);
-  
-}
-
-.search-button:active {
-  transform: scale(0.95);
-  box-shadow: 0 0 8px 4px rgba(0, 0, 0, 0.1); /* Consistent shadow effect */
-}
-
-.search-bar {
-  position: absolute;
-  top: -30px;
-  height: 50px;
-  margin-left: 10px;
-  padding: 0 15px;
-  border: none;
-  border-radius: 25px;
-  background-color: #F1EBE5;
-  color: #5c4033;
-  font-size: 16px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  transition: box-shadow 0.3s ease;
- 
-}
-
-.search-bar:hover, .search-bar:focus {
-  outline: none;
-  box-shadow: 0 6px 12px rgba(166, 124, 104, 0.3);
-}
-
-.search-container {
-  gap: 10px;
-}
-
-
-
-/* Responsive adjustments */
-@media(width:768){
-  .search-container.active .search-bar {
-    width: 140px;
-   left:-80px;
-
-  }
-}
-@media(max-width:768px){
-  .search-container.active .search-bar {
-    width: 140px;
-   left:-80px
-  }
-  .search-button{
-    left:-105px
-  }
-}
-@media (max-width: 1015px)and (min-width:768px) {
-  
-  .search-container.active .search-bar {
-    width: 170px;
-   left:300px ;
-     top:10px;
-  }
-  .search-button{
-    left:290px
-  }
-  .nav-link:hover::after {
-    width:0;
-  }
-}
-@media(min-width:1015px){
-  .search-container.active .search-bar {
-    width: 250px;
-   left:390px;
-  }
-  .search-button{
-    left: 300px;
-  }
-
-}
-
-@media(min-width:1290px){
-  .search-container.active .search-bar {
-    width: 490px;
-   
-  }
-  .search-button{
-    left: 360px;
-  }
-
-}
-@media(min-width:1500px){
-  .search-container.active .search-bar {
-    width: 490px;
-   left:360px;
-  }
-  .search-button{
-   left: 320px;
-  }
-
 }
