+++ src/components/NavBar.js
@@ -1,15 +1,12 @@
 //import
 import { Link } from "react-router-dom";
 import "./NavBar.css";
-import React, { useState } from "react";
+import React from "react";
 import logo from './img/logo.png';
+import SearchBar from "./SearchBar";
 
 const NavBar = ({ BrandName, i2, i3 }) => {
-  const [isSearching, setIsSearching] = useState(false);
 
-  const handleSearchToggle = () => {
-    setIsSearching((prev) => !prev);
-  };
 
   return (
     <div className="container">
@@ -20,22 +17,7 @@ const NavBar = ({ BrandName, i2, i3 }) => {
         <img src={logo} alt="Logo" className="logo" /> 
           {BrandName}
         </Link>
-        <div className={`search-container ${isSearching ? "active" : ""}`}>
-          <button className="search-button" onClick={handleSearchToggle}>
-            {isSearching ? (
-              <i className="fas fa-times"></i>
-            ) : (
-              <i className="fas fa-search"></i>
-            )}
-          </button>
-          {isSearching && (
-            <input
-              type="text"
-              className="search-bar"
-              placeholder="Search..."
-            />
-          )}
-        </div>
+       <SearchBar/>
 
 
         <button
