+++ src/components/SearchBar.js
@@ -1,22 +1,30 @@
-import React from "react";
+import React, { useState } from "react";
 import "./SearchBar.css";
 
 const SearchBar = () => {
+  const [isSearching, setIsSearching] = useState(false);
+    
+  const handleSearchToggle = () => {
+    setIsSearching((prev) => !prev);
+  };
   return (
-    <div className="grp-6-container container-fluid">
-      <form>
-        <div className="input-group">
-          <input
-            className="form-control"
-            placeholder="Search for Destination, restaurant ..."
-            type="text"
-          />
-          <button className="btn btn-dark" type="submit">
-            <i className="fa-solid fa-magnifying-glass"></i>
-          </button>
-        </div>
-      </form>
-    </div>
+    <div className={`search-container ${isSearching ? "active" : ""}`}>
+    <button className="search-button" onClick={handleSearchToggle}>
+      {isSearching ? (
+        <i className="fas fa-times"></i>
+      ) : (
+        <i className="fas fa-search"></i>
+      )}
+    </button>
+    {isSearching && (
+      <input
+        type="text"
+        className="search-bar"
+        placeholder="Search..."
+      />
+    )}
+  </div>
+    
   );
 };
